requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.86.0
greenlet==2.0.0.post0
h11==0.14.0
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.4
jupyter_core==4.11.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
stack-data==0.6.0
starlette==0.20.4
tornado==6.2
traitlets==5.5.0
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
